Prevent potential bootloader brick and/or untrusted code run
The first instruction in MCU's flash block 0 are a "GOTO bootloader_address"
instruction(4 bytes). This sector gets erased as preparation for
flashing procedures.

The instruction is saved to RAM during flash block 0 erase and then flashed back
immediately after block 0 erase.
It can happen though for a MCU reset or power cut to occur in the time between
sector 0 erase and "GOTO bootloader_address" rewrite from RAM to flash.

In this case, when MCU will start, instead of jumping to bootloader code, it
will slide through NOP(0xff opcode) instructions into whatever code is valid
in the second flash block. Thus bootloader will not start and code that
was not validated for signature will be executed.

To prevent this problem, this commit adds logic to flash the second block
of flash with "GOTO bootloader_address" instruction. This way, after block 0
is erased and filled with NOPs, if a reset or power cut occurs immediately,
the MCU will slide through block 0 into the "GOTO bootloader_address" that is placed at
start of flash block 1.

During normal operation, after successfully erasing flash sector 0 and writing back
the bootloader jump instruction, block 1 will be erased and will be prepared for user
code write

Signed-off-by: Sebastian Panceac <spanceac@gmail.com>