Have a single exit point from bootloader
Previous to this commit, the code was executing a reset instruction
after flashing was done directly in the message handler.
The code now returns from the message handler until the main()
function and has a single exit point at the end of main().

This helps with following the code logic

Signed-off-by: Sebastian Panceac <spanceac@gmail.com>